Rename Types to Fields
