+++ src/api/handler.ts
@@ -12,7 +12,7 @@ import AuthedTaskModel from './db/models/Task.js';
 import Query from './resolvers/Query.js';
 import Mutation from './resolvers/Mutation.js';
 import Scalars from './resolvers/Scalars.js';
-import Types from './resolvers/Types.js';
+import Fields from './resolvers/Fields.js';
 import typeDefs from './type-defs/index.js';
 import { Config, getConfig } from '../config/config.js';
 import { connectToDatabase } from './db/connect.js';
@@ -24,7 +24,7 @@ const resolvers = {
   Query,
   Mutation,
   ...Scalars,
-  ...Types,
+  ...Fields,
 };
 
 const corsMiddleware = async () => {
